Move adapater to plugins api
